zjazd09/1_sobota/invoice_generator_project/pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "invoice-generator-merito-2025"
version = "0.0.1"

dependencies = [
    "fpdf2",
    "pydantic"
]

[tool.hatch.build.targets.wheel]
packages = ["invoice_generator"]